==> 1)Raw_Data.ipynb <==
import json
import pandas as pd

def extract_transform_load(kaggle_metadata_path, movielens_rating_data_path, wikipedia_data_path):   
    
    kaggle_metadata = pd.read_csv(kaggle_metadata_path, low_memory=False) # Create dataframe from local file path
    kaggle_df = pd.DataFrame(kaggle_metadata)
    
    movielens_rating_data = pd.read_csv(movielens_rating_data_path, low_memory=False)
    movielens_rating_df = pd.DataFrame(movielens_rating_data)

    with open(wikipedia_data_path, mode='r') as file: # Open/load json data
        wiki_movies_raw = json.load(file)

    wiki_movies_df = pd.DataFrame(wiki_movies_raw)

    return kaggle_df, movielens_rating_df, wiki_movies_df

file_dir = "archive" # Define paths
kaggle_metadata_path = f'{file_dir}/movies_metadata.csv'
movielens_rating_data_path = f'{file_dir}/ratings.csv'
wikipedia_data_path = f'{file_dir}/wikipedia-movies.json'

kaggle_df, movielens_rating_df, wiki_movies_df = extract_transform_load(kaggle_metadata_path, 
                                                                        movielens_rating_data_path, 
                                                                        wikipedia_data_path)

output_data_file = f'{file_dir}/raw_kaggle_df.csv'
kaggle_df.to_csv(output_data_file, index_label="Movie number")
kaggle_df

output_data_file = f'{file_dir}/raw_movielens_rating_df.csv'
movielens_rating_df.to_csv(output_data_file, index_label="Movie number")
movielens_rating_df

output_data_file = f'{file_dir}/raw_wiki_movies_df.csv'
wiki_movies_df.to_csv(output_data_file, index_label="Movie number")


wiki_movies_df
