==> 2)Json_Cleaning.ipynb <==
import json
import re
import numpy as np
import pandas as pd

file_dir = "archive" # Define paths
wikipedia_data_raw_path = f'{file_dir}/wikipedia-movies.json'

def Cleaner(wikipedia_data_raw_path):
    
    with open(wikipedia_data_raw_path, mode='r') as file: # Load json data
        wiki_movies_raw = json.load(file)
         
    wiki_movies_raw = [movie for movie in wiki_movies_raw # Filter out tv shows
                   if ('Director' in movie or 'Directed by' in movie)
                   and 'imdb_link' in movie
                   and 'No. of episodes' not in movie]
    
    # Add clean_movie function
    def clean_movie(movie):
        movie = dict(movie) # Create a non-destructive copy
        alt_titles = {} # Combine alternate titles into one list
        
        for key in ['Also known as','Arabic','Cantonese','Chinese','French',
                'Hangul','Hebrew','Hepburn','Japanese','Literally',
                'Mandarin','McCune-Reischauer','Original title','Polish',
                'Revised Romanization','Romanized','Russian',
                'Simplified','Traditional','Yiddish']:
            if key in movie:
                alt_titles[key] = movie[key]
                movie.pop(key)
        if len(alt_titles) > 0:
            movie['alt_titles'] = alt_titles
            
        def change_column_name(old_name, new_name): # Merge column names
            if old_name in movie:
                movie[new_name] = movie.pop(old_name)
        change_column_name('Adaptation by', 'Writer(s)')
        change_column_name('Country of origin', 'Country')
        change_column_name('Directed by', 'Director')
        change_column_name('Distributed by', 'Distributor')
        change_column_name('Edited by', 'Editor(s)')
        change_column_name('Length', 'Running time')
        change_column_name('Original release', 'Release date')
        change_column_name('Music by', 'Composer(s)')
        change_column_name('Produced by', 'Producer(s)')
        change_column_name('Producer', 'Producer(s)')
        change_column_name('Productioncompanies ', 'Production company(s)')
        change_column_name('Productioncompany ', 'Production company(s)')
        change_column_name('Released', 'Release Date')
        change_column_name('Release Date', 'Release date')
        change_column_name('Screen story by', 'Writer(s)')
        change_column_name('Screenplay by', 'Writer(s)')
        change_column_name('Story by', 'Writer(s)')
        change_column_name('Theme music composer', 'Composer(s)')
        change_column_name('Written by', 'Writer(s)')
        
        return movie  
    clean_movies = [clean_movie(movie) for movie in wiki_movies_raw] # clean json dict names
    wiki_movies_df = pd.DataFrame(clean_movies) # convert to a dataframe

    # Drop imdb_id duplicates and filter based on Regex
    try:
        wiki_movies_df['imdb_id'] = wiki_movies_df['imdb_link'].str.extract(r'(tt\d{7})')
        wiki_movies_df = wiki_movies_df.drop_duplicates(subset=['imdb_id'])
        
    except:
        print('unavalible for extraction')
    
    #  Remove nulls
    wiki_columns_to_keep = [column for column in wiki_movies_df.columns 
                            if wiki_movies_df[column].isnull().sum() < len(wiki_movies_df) * 0.9]
    wiki_movies_df = wiki_movies_df[wiki_columns_to_keep]
    
    # Store columns to be cleaned in variables, drop na, convert list items to strings
    box_office = wiki_movies_df['Box office'].dropna().apply(lambda x: ' '.join(x) if type(x) == list else x)
    budget = wiki_movies_df['Budget'].dropna().map(lambda x: ' '.join(x) if type(x) == list else x)
    release_date = wiki_movies_df['Release date'].dropna().apply(lambda x: ' '.join(x) if type(x) == list else x)
    running_time = wiki_movies_df['Running time'].dropna().apply(lambda x: ' '.join(x) if type(x) == list else x)

    form_one = r'\$\s*\d+\.?\d*\s*[mb]illi?on' # Regex, match the six elements of "form_one" in box office data.
    form_two = r'\$\s*\d{1,3}(?:[,\.]\d{3})+(?!\s[mb]illion)' # three elements of "form_two" of the box office data.
    date_form_one = r'(?:January|February|March|April|May|June|July|August|September|October|November|December)\s[123]?\d,\s\d{4}'
    date_form_three = r'(?:January|February|March|April|May|June|July|August|September|October|November|December)\s\d{4}'
    date_form_two, date_form_four = r'\d{4}.[01]\d.[0123]\d', r'\d{4}'
    
    box_office = box_office.str.replace(r'\$.*[-—–](?![a-z])', '$', regex=True)
    budget = budget.str.replace(r'\$.*[-—–](?![a-z])', '$', regex=True)
    budget = budget.str.replace(r'\[\d+\]\s*', '')
        
    box_office = box_office.str.extract(f'({form_one}|{form_two})', flags=re.IGNORECASE)
    matches_form_one = budget.str.contains(form_one, flags=re.IGNORECASE, na=False)
    matches_form_two = budget.str.contains(form_two, flags=re.IGNORECASE, na=False)
    budget[~matches_form_one & ~matches_form_two]
    running_time_extract = running_time.str.extract(r'(\d+)\s*ho?u?r?s?\s*(\d*)|(\d+)\s*m')
    running_time_extract = running_time_extract.apply(lambda col: pd.to_numeric(col, errors='coerce')).fillna(0)    
    
    # Add the parse_dollars function.
    def parse_dollars(s): 
        if type(s) != str:
            return np.nan
        
        if re.match(r'\$\s*\d+\.?\d*\s*milli?on', s, flags=re.IGNORECASE):
            s = re.sub('\$|\s|[a-zA-Z]','', s) # remove dollar sign and "million"
            value = float(s) * 10**6 # convert to float and multiply by a million
            
            return value 

        elif re.match(r'\$\s*\d+\.?\d*\s*billi?on', s, flags=re.IGNORECASE):
            s = re.sub('\$|\s|[a-zA-Z]','', s) # remove dollar sign and "billion"
            value = float(s) * 10**9 # convert to float and multiply by a billion

            return value 

        elif re.match(r'\$\s*\d{1,3}(?:[,\.]\d{3})+(?!\s[mb]illion)', s, flags=re.IGNORECASE):
            s = re.sub('\$|,','', s) # remove dollar sign and commas
            value = float(s)# convert to float

            return value

        else:
            return np.nan
    
    wiki_movies_df['box_office'] = box_office[0].apply(parse_dollars)
    wiki_movies_df['budget'] = budget.str.extract(f'({form_one}|{form_two})', 
                                                  flags=re.IGNORECASE)[0].apply(parse_dollars)
    wiki_movies_df['release_date'] = pd.to_datetime(release_date.str.extract(
            f'({date_form_one}|{date_form_two}|{date_form_three}|{date_form_four})')[0], infer_datetime_format=True)
    wiki_movies_df['running_time'] = running_time_extract.apply(lambda row: 
                                                                row[0]*60 + row[1] if row[2] == 0 else row[2], axis=1)

    wiki_movies_df.drop('Box office', axis=1, inplace=True)
    wiki_movies_df.drop('Budget', axis=1, inplace=True)
    wiki_movies_df.drop('Release date', axis=1, inplace=True)
    wiki_movies_df.drop('Running time', axis=1, inplace=True)

    return wiki_movies_df 


wiki_movies_df = Cleaner(wikipedia_data_raw_path)
wiki_movies_df.to_csv(f'{file_dir}/clean_wiki_movies_df.csv', index_label= False)
wiki_movies_df

# Different from test because cleaned og columns dropped. Can reference og in raw_wikipedia_movies_df.csv
wiki_movies_df.columns.to_list()
