==> 3)CSV_Cleaning.ipynb <==
import pandas as pd

file_dir = "archive" # Define paths
kaggle_metadata_path = f'{file_dir}/raw_kaggle_df.csv'
movielens_rating_data_path = f'{file_dir}/raw_movielens_rating_df.csv'
wiki_movies_path = f'{file_dir}/clean_wiki_movies_df.csv'

def Cleanr(kaggle_metadata_path, movielens_rating_data_path, wiki_movies_path):
    
    kaggle_metadata = pd.read_csv(kaggle_metadata_path, low_memory=False) # Read csv data
    movielens_rating_data = pd.read_csv(movielens_rating_data_path, low_memory=False)
    wiki_movies_data = pd.read_csv(wiki_movies_path, low_memory=False)

    kaggle_metadata = pd.DataFrame(kaggle_metadata) # Create dataframe from local file path
    movielens_rating_df = pd.DataFrame(movielens_rating_data)
    wiki_movies_df = pd.DataFrame(wiki_movies_data)
    
    # Clean the Kaggle & Movie lengths metadata
    kaggle_metadata = kaggle_metadata[kaggle_metadata['adult'] == 'False'].drop('adult',axis='columns')
    kaggle_metadata['video'] = kaggle_metadata['video'] == 'True'
    kaggle_metadata['budget'] = kaggle_metadata['budget'].astype(int)
    kaggle_metadata['id'] = pd.to_numeric(kaggle_metadata['id'], errors='raise')
    kaggle_metadata['popularity'] = pd.to_numeric(kaggle_metadata['popularity'], errors='raise')
    kaggle_metadata['release_date'] = pd.to_datetime(kaggle_metadata['release_date'])
    movielens_rating_df['timestamp'] = pd.to_datetime(movielens_rating_df['timestamp'], unit='s')
    movielens_rating_df = movielens_rating_df.rename(columns={"movieId": "id"})
    
    # Merged the wiki movies and kaggle movies DataFrames 
    movies_df = pd.merge(wiki_movies_df, kaggle_metadata, how = 'inner', on='imdb_id', suffixes=['_wiki','_kaggle'])

    # Drop unnecessary columns from merge
    movies_df.drop(columns=['title_wiki','release_date_wiki','Language','Production company(s)'], inplace=True)
    
    # Fill in the missing Kaggle data
    def fill_missing_kaggle_data(df, kaggle_column, wiki_column):
        df[kaggle_column] = df.apply(
            lambda row: row[wiki_column] if row[kaggle_column] == 0 else row[kaggle_column], axis=1)
        df.drop(columns=wiki_column, inplace=True)   
       
    # Call the function with DataFrame and columns as the arguments
    fill_missing_kaggle_data(movies_df, 'runtime', 'running_time')
    fill_missing_kaggle_data(movies_df, 'budget_kaggle', 'budget_wiki')
    fill_missing_kaggle_data(movies_df, 'revenue', 'box_office')
    
    # Filter movies DataFrame for specific columns.
    for col in movies_df.columns:
        lists_to_tuples = lambda x: tuple(x) if type(x) == list else x
        value_counts = movies_df[col].apply(lists_to_tuples).value_counts(dropna=False)
        num_values = len(value_counts)
        if num_values == 1:
            movies_df.drop(columns=['video'], inplace=True)
            
    movielens_rating_df = movielens_rating_df.drop_duplicates(subset=['id'])
     
    # Merge the ratings DataFrame.
    movies_df = pd.merge(movies_df, movielens_rating_df, how='left', on='id', suffixes=['_wikis','_ratings'])  
    
    # Rename and Rearrange new columns         
    movies_df = movies_df.loc[:, ['imdb_id','id','title_kaggle','original_title','tagline','belongs_to_collection',
                                  'url','imdb_link','runtime','budget_kaggle','revenue','release_date_kaggle',
                                  'popularity','vote_average','vote_count','genres','original_language','overview',
                                  'spoken_languages','Country','production_companies','production_countries',
                                  'Distributor','Producer(s)','Director','Starring','Cinematography','Editor(s)',
                                  'Writer(s)','Composer(s)','Based on']]

    movies_df.rename({'id':'kaggle_id','title_kaggle':'title','url':'wikipedia_url','budget_kaggle':'budget',
                      'release_date_kaggle':'release_date','Country':'country','Distributor':'distributor',
                      'Producer(s)':'producers','Director':'director','Starring':'starring',
                      'Cinematography':'cinematography','Editor(s)':'editors','Writer(s)':'writers',
                      'Composer(s)':'composers','Based on':'based_on'}, axis='columns', inplace=True)

    return movies_df


movies_df = Cleanr(kaggle_metadata_path, movielens_rating_data_path, wiki_movies_path)
movies_df.to_csv(f'{file_dir}/cleaned_merged_movies_df.csv', index_label= False)
movies_df
