==> 4)Create_Database.ipynb <==
import time
import psycopg2
import pandas as pd
from config import db_password
from sqlalchemy import create_engine

file_dir = "archive" # Define paths
db_string = f"postgresql://postgres:{db_password}@127.0.0.1:5432/movie_data"
merged = f'{file_dir}/cleaned_merged_movies_df.csv'
merged_movies_df = pd.read_csv(merged, low_memory=False)

engine = create_engine(db_string)
merged_movies_df.to_sql(name='wiki_movies', con=engine)

rows_imported = 0 # initialize count
start_time = time.time()
for data in pd.read_csv(f'{file_dir}/ratings.csv', chunksize=1000000):
    
    print(f'importing rows {rows_imported} to {rows_imported + len(data)}...', end='') # print rows imported
    data.to_sql(name='ratings', con=engine, if_exists='append')
    rows_imported += len(data) # increment the number of rows imported by the size of 'data'
    
    print(f'Done. {time.time() - start_time} total seconds elapsed') # print that the rows have finished importing
